restructure code into dedicated packages
