+++ .github/workflows/run-all-tests.yaml
@@ -35,4 +35,4 @@ jobs:
       - name: Install Dependencies
         run: npm install
       - name: Run all tests
-        run: npm run test
+        run: npm run test:all
