+++ .prettierignore
@@ -1,4 +1,5 @@
 dist
 coverage
 node_modules
+packages/fuzzer/build
 .idea
