+++ package.json
@@ -1,13 +1,13 @@
 {
 	"name": "fuzzyeagle",
 	"version": "0.0.0",
-	"description": "Coverage-guided, in-process fuzzing for the Node.js",
+	"description": "Coverage-guided, in-process fuzzing for Node.js",
 	"homepage": "https://github.com/CodeIntelligenceTesting/fuzzyEagle#readme",
 	"keywords": [
 		"coverage-guided",
 		"fuzzing"
 	],
-	"author": "",
+	"author": "Code Intelligence",
 	"license": "Apache-2.0",
 	"bugs": {
 		"url": "https://github.com/CodeIntelligenceTesting/fuzzyEagle/issues"
@@ -20,8 +20,9 @@
 	"scripts": {
 		"compile": "tsc -p .",
 		"build": "tsc",
-		"test": "jest && cd fuzzer && npm install && npm test",
+		"test": "jest",
 		"test:watch": "jest --watch",
+		"test:all": "./scripts/test.sh",
 		"lint": "eslint .",
 		"lint:fix": "npm run lint -- --fix",
 		"prettier": "prettier --check .",
@@ -32,16 +33,8 @@
 	"lint-staged": {
 		"**/*": "prettier --write --ignore-unknown"
 	},
-	"dependencies": {
-		"istanbul-lib-hook": "^3.0.0",
-		"yargs": "^17.5.1"
-	},
 	"devDependencies": {
-		"@babel/core": "^7.18.6",
-		"@types/babel__core": "^7.1.19",
-		"@types/istanbul-lib-hook": "^2.0.1",
 		"@types/jest": "^28.1.4",
-		"@types/node": "^18.0.0",
 		"@types/yargs": "^17.0.10",
 		"@typescript-eslint/eslint-plugin": "^5.30.0",
 		"@typescript-eslint/parser": "^5.30.0",
@@ -54,5 +47,8 @@
 		"prettier": "2.7.1",
 		"ts-jest": "^28.0.5",
 		"typescript": "^4.7.4"
-	}
+	},
+	"workspaces": [
+		"packages/*"
+	]
 }
