+++ packages/fuzzer/README.md
@@ -14,10 +14,10 @@ users want, so that their JS fuzz target can run in its normal environment.
 
 The project can be built with `npm run compile` (which is incremental after the
 first build); a subsequent `npm test` makes sure that the plugin loads cleanly.
-Binaries can be prebuilt with `npm run prebuild` and uploaded with `npm run
-upload`.
+Binaries can be prebuilt with `npm run prebuild` and uploaded with `npm run upload`.
 
 Internally, the build system uses several steps:
+
 1. `package.json` defines the `npm` scripts.
 2. Several of them use `prebuild` or `prebuild-install`; together, those two
    tools implement a binary cache via GitHub releases, so that users don't have
@@ -29,8 +29,7 @@ Internally, the build system uses several steps:
 5. In our CMake configuration, we set up compiler-rt as an external project;
    CMake fetches and builds it before compiling our own code against it.
 
-To debug build issues, it's often useful to start with a plain `cmake-js
-compile` or `cmake-js recompile`, which just invokes CMake with a few extra
+To debug build issues, it's often useful to start with a plain `cmake-js compile` or `cmake-js recompile`, which just invokes CMake with a few extra
 arguments that help it to find the Node headers and such.
 
 When working on the plugin's C++ code, you may want to use a language server
