+++ packages/fuzzer/package.json
@@ -1,0 +1,33 @@
+{
+	"name": "@fuzzy-eagle/fuzzer",
+	"description": "A libfuzzer-based fuzzer for Node",
+	"author": "Code Intelligence",
+	"version": "0.0.0",
+	"license": "Apache-2.0",
+	"main": "fuzzer.js",
+	"scripts": {
+		"test": "node test_fuzzer.js",
+		"prebuild": "prebuild --runtime napi --backend cmake-js --all --strip --verbose",
+		"upload": "prebuild --runtime napi --backend cmake-js --all --strip --upload ${GITHUB_TOKEN}",
+		"install": "prebuild-install --runtime napi || cmake-js rebuild",
+		"compile": "cmake-js compile"
+	},
+	"dependencies": {
+		"bindings": "^1.5.0",
+		"cmake-js": "^6.3.2",
+		"node-addon-api": "^5.0.0",
+		"prebuild-install": "^7.1.1"
+	},
+	"binary": {
+		"napi_versions": [
+			3
+		]
+	},
+	"devDependencies": {
+		"prebuild": "^11.0.3"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/CodeIntelligenceTesting/fuzzyEagle.git"
+	}
+}
