+++ packages/fuzzer/test_fuzzer.js
