+++ packages/instrumentor/package.json
@@ -1,0 +1,17 @@
+{
+	"name": "@fuzzy-eagle/instrumentor",
+	"version": "0.0.0",
+	"description": "Instrumentor",
+	"author": "Code Intelligence",
+	"license": "Apache-2.0",
+	"dependencies": {
+		"@babel/core": "^7.18.6",
+		"istanbul-lib-hook": "^3.0.0"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.18.6",
+		"@types/babel__core": "^7.1.19",
+		"@types/istanbul-lib-hook": "^2.0.1",
+		"@types/node": "^18.0.0"
+	}
+}
